Add a test to check bad use of application owned memory
Add the test vulkan_pnext_chain that checks the use of application
owned memory passed through pNext chains. A bug has been found in
GFXReconstruct where it uses application owned pointer after the
vulkan call that gave the pointer, thus with application having the
possibility of freeing the underlying memory. This unsafe use causes
a SEGFAULT. This is a minimal repro of this bug.