Add another test case for silly app API usage
This one copies the feature and extension lists from the
offered features and extensions into the required ones.
It is really dumb but a lot of apps do this for some
reason and it is perfectly valid usage of the API.

For now it only copies extensions. Features coming later.